--- spec/lib/gems_bond/printer_spec.rb
@@ -1,5 +1,0 @@
-# frozen_string_literal: true
-
-require "gems_bond/printer"
-
-RSpec.describe GemsBond::Printer do; end # rubocop:disable RSpec/EmptyExampleGroup, Style/BlockDelimiters
