+++ spec/lib/gems_bond/printers/html_spec.rb
@@ -1,0 +1,5 @@
+# frozen_string_literal: true
+
+require "gems_bond/printers/html"
+
+RSpec.describe GemsBond::Printers::HTML do; end # rubocop:disable RSpec/EmptyExampleGroup, Style/BlockDelimiters
